requirements.txt
# ============================================
# Core Django
# ============================================
Django==5.0.10
django-environ==0.11.2
psycopg2-binary==2.9.9
setuptools>=80.0.0

# ============================================
# Storage & Files
# ============================================
Pillow==10.1.0

# ============================================
# Forms & UI
# ============================================
django-crispy-forms==2.3
crispy-bootstrap5==2025.6
django-widget-tweaks==1.5.0

# ============================================
# Utilities
# ============================================
python-dotenv==1.0.0
django-extensions==3.2.3
python-dateutil==2.8.2
django-countries==7.5.1

# ============================================
# REST API
# ============================================
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
django-ratelimit==4.1.0

# ============================================
# Security
# ============================================
django-cors-headers==4.3.1

# ============================================
# Development Tools
# ============================================
black==23.12.1
flake8==7.0.0
django-debug-toolbar==4.2.0
ipython>=8.20,<9

# ============================================
# Testing
# ============================================
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0
